+++ api/trusted_setup.go
@@ -17,10 +17,10 @@ import (
 // - Check that the points are in the correct subgroup.
 // - Check that setupG1Lagrange is the lagrange version of setupG1.
 //
-// Note: There is an embedded (via a //go:embed - compiler instruction) setup testKzgSetupStr, to which we do check those properties
-// in a test function.
+// Note: There is an embedded (via a //go:embed - compiler instruction) setup
+// testKzgSetupStr, to which we do check those properties in a test function.
 
-// JSONTrustedSetup is a struc used for JSON-(de)-serializing the trusted setup from/to JSON format.
+// JSONTrustedSetup is a struct used for serializing the trusted setup from/to JSON format.
 //
 // The intended use-case is that library users store the trusted setup in a JSON file and we provide such a file
 // as part of the package.
@@ -50,9 +50,6 @@ var (
 // lagrange version of G1 points
 // - All elements are in the correct subgroup
 // - Lagrange G1 points are obtained by doing an IFFT of monomial G1 points
-//
-// Note that in a honestly generated trusted setup, the G1 points are related.
-// CheckTrustedSetupIsWellFormed does not (and, lacking the secret used to construct the SRS, cannot) verify this.
 func CheckTrustedSetupIsWellFormed(trustedSetup *JSONTrustedSetup) error {
 
 	if len(trustedSetup.SetupG1) != len(trustedSetup.SetupG1Lagrange) {
@@ -101,8 +98,9 @@ func CheckTrustedSetupIsWellFormed(trustedSetup *JSONTrustedSetup) error {
 	return nil
 }
 
-// parseTrustedSetup parses the trusted setup in JSONTrustedSetup format (which essentially holds strings)into corresponding group elements.
-// Elements are assumed to be trusted.
+// parseTrustedSetup parses the trusted setup in `JSONTrustedSetup` format
+// which contains hex encoded strings to corresponding group elements.
+// Elements are assumed to be well-formed.
 func parseTrustedSetup(trustedSetup *JSONTrustedSetup) (bls12381.G1Affine, []bls12381.G1Affine, []bls12381.G2Affine, error) {
 	// Take the generator point from the monomial SRS
 	if len(trustedSetup.SetupG1) < 1 {
@@ -126,9 +124,10 @@ func parseTrustedSetup(trustedSetup *JSONTrustedSetup) (bls12381.G1Affine, []bls
 	return genG1, setupLagrangeG1Points, g2Points, nil
 }
 
-// parseG1PointNoSubgroupCheck parses a hex-string (without 0x prefix or the like) into a G1 point.
+// parseG1PointNoSubgroupCheck parses a hex-string (without 0x prefix) into a G1 point.
 //
-// This function performs no (expensive) subgroup checks, and should only be used for trusted inputs.
+// This function performs no (expensive) subgroup checks, and should only be used
+// for trusted inputs.
 func parseG1PointNoSubgroupCheck(hexString string) (bls12381.G1Affine, error) {
 	byts, err := hex.DecodeString(hexString)
 	if err != nil {
@@ -143,9 +142,10 @@ func parseG1PointNoSubgroupCheck(hexString string) (bls12381.G1Affine, error) {
 
 }
 
-// parseG2PointNoSubgroupCheck parses a hex-string (without 0x prefix or the like) into a G2 point.
+// parseG2PointNoSubgroupCheck parses a hex-string (without 0x prefix) into a G2 point.
 //
-// This function performs no (expensive) subgroup checks, and should only be used for trusted inputs.
+// This function performs no (expensive) subgroup checks, and should only be used
+// for trusted inputs.
 func parseG2PointNoSubgroupCheck(hexString string) (bls12381.G2Affine, error) {
 	byts, err := hex.DecodeString(hexString)
 	if err != nil {
@@ -159,11 +159,14 @@ func parseG2PointNoSubgroupCheck(hexString string) (bls12381.G2Affine, error) {
 	return point, d.Decode(&point)
 }
 
-// parseG1PointsNoSubgroupCheck parses a slice hex-string (without 0x prefix or the like) into a slice of G1 points.
+// parseG1PointsNoSubgroupCheck parses a slice hex-string (without 0x prefix) into a
+// slice of G1 points.
 //
-// This is essentially a parallelized version of calling [parseG1PointNoSubgroupCheck] on each element of the slice individually.
+// This is essentially a parallelized version of calling [parseG1PointNoSubgroupCheck]
+// on each element of the slice individually.
 //
-// This function performs no (expensive) subgroup checks, and should only be used for trusted inputs.
+// This function performs no (expensive) subgroup checks, and should only be used
+// for trusted inputs.
 func parseG1PointsNoSubgroupCheck(hexStrings []string) ([]bls12381.G1Affine, error) {
 	numG1 := len(hexStrings)
 	g1Points := make([]bls12381.G1Affine, numG1)
@@ -185,11 +188,14 @@ func parseG1PointsNoSubgroupCheck(hexStrings []string) ([]bls12381.G1Affine, err
 	return g1Points, nil
 }
 
-// parseG2PointsNoSubgroupCheck parses a slice hex-string (without 0x prefix or the like) into a slice of G2 points.
+// parseG2PointsNoSubgroupCheck parses a slice hex-string (without 0x prefix) into a
+// slice of G2 points.
 //
-// This is essentially a parallelized version of calling [parseG2PointNoSubgroupCheck] on each element of the slice individually.
+// This is essentially a parallelized version of calling [parseG2PointNoSubgroupCheck]
+// on each element of the slice individually.
 //
-// This function performs no (expensive) subgroup checks, and should only be used for trusted inputs.
+// This function performs no (expensive) subgroup checks, and should only be used
+// for trusted inputs.
 func parseG2PointsNoSubgroupCheck(hexStrings []string) ([]bls12381.G2Affine, error) {
 	numG2 := len(hexStrings)
 	g2Points := make([]bls12381.G2Affine, numG2)
