+++ internal/kzg/srs.go
@@ -1,8 +1,6 @@
 package kzg
 
 import (
-	"math/big"
-
 	bls12381 "github.com/consensys/gnark-crypto/ecc/bls12-381"
 	"github.com/consensys/gnark-crypto/ecc/bls12-381/fr"
 	"github.com/crate-crypto/go-proto-danksharding-crypto/internal/multiexp"
@@ -30,7 +28,7 @@ type CommitKey struct {
 	G1 []bls12381.G1Affine
 }
 
-// ReversePoints Applies the bit reversal permutation
+// ReversePoints applies the bit reversal permutation
 // to the G1 points stored inside the CommitKey c.
 func (c *CommitKey) ReversePoints() {
 	bitReverse(c.G1)
@@ -46,74 +44,6 @@ func (c *CommitKey) ReversePoints() {
 type SRS struct {
 	CommitKey  CommitKey
 	OpeningKey OpeningKey
-}
-
-// newLagrangeSRSInsecure creates a new SRS object with the secret `bAlpha`. The resulting SRS uses the Lagrange basis
-//
-// This method should not be used in production because this is not a trusted setup,
-// as at least one person (the caller, for a start) knows what `bAlpha`is.
-func newLagrangeSRSInsecure(domain Domain, bAlpha *big.Int) (*SRS, error) {
-	return newSRSInsecure(domain, bAlpha, true)
-}
-
-// newSRSInsecure creates a new SRS object with the secret `bAlpha`.
-// convertToLagrange controls whether the result is in monomial or Lagrange basis.
-//
-// This method should not be used in production because this is not a trusted setup,
-// as at least one person (the caller, for a start) knows what `bAlpha`is.
-func newSRSInsecure(domain Domain, bAlpha *big.Int, convertToLagrange bool) (*SRS, error) {
-	srs, err := newMonomialSRSInsecure_uint64(domain.Cardinality, bAlpha)
-	if err != nil {
-		return nil, err
-	}
-
-	if convertToLagrange {
-		// Convert SRS from monomial form to lagrange form
-		lagrangeG1 := domain.IfftG1(srs.CommitKey.G1)
-		srs.CommitKey.G1 = lagrangeG1
-	}
-
-	return srs, nil
-}
-
-// newMonomialSRSInsecure_uint64 creates a new SRS object with the secret `bAlpha` in monomial basis.
-//
-// This is only used for testing. As with the other new<foo>SRSInsecure functions, its is insecure, because we provide the secret scalar as input.
-// Consequently, this method should also never be used in production. Note that the function is named _size, because we provide
-// the size argument as a uint64 rather than a Domain. A newMonomialSRSInsecure functions taking a Domain as input to match the other functions is defined in the
-// testing code.
-//
-// Copied from [gnark-crypto](https://github.com/ConsenSys/gnark-crypto/blob/8f7ca09273c24ed9465043566906cbecf5dcee91/ecc/bls12-381/fr/kzg/kzg.go#L65)
-func newMonomialSRSInsecure_uint64(size uint64, bAlpha *big.Int) (*SRS, error) {
-	if size < 2 {
-		return nil, ErrMinSRSSize
-	}
-
-	var commitKey CommitKey
-	var openKey OpeningKey
-	commitKey.G1 = make([]bls12381.G1Affine, size)
-
-	var alpha fr.Element
-	alpha.SetBigInt(bAlpha)
-
-	_, _, gen1Aff, gen2Aff := bls12381.Generators()
-	commitKey.G1[0] = gen1Aff
-	openKey.GenG1 = gen1Aff
-	openKey.GenG2 = gen2Aff
-	openKey.AlphaG2.ScalarMultiplication(&gen2Aff, bAlpha)
-
-	alphas := make([]fr.Element, size-1)
-	alphas[0] = alpha
-	for i := 1; i < len(alphas); i++ {
-		alphas[i].Mul(&alphas[i-1], &alpha)
-	}
-	g1s := bls12381.BatchScalarMultiplicationG1(&gen1Aff, alphas)
-	copy(commitKey.G1[1:], g1s)
-
-	return &SRS{
-		CommitKey:  commitKey,
-		OpeningKey: openKey,
-	}, nil
 }
 
 // Commit commits to a polynomial using a multi exponentiation with the
